app/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
openai-whisper==20231117
pydantic==2.9.2
python-multipart==0.0.9
torch
torchaudio
numpy
